feat: complete exercise 4.21

+++ part4/bloglist/controllers/blogs.js
@@ -32,8 +32,19 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const requestUser = await User.findById(decodedToken.id)
+  const blogToDelete = await Blog.findById(request.params.id)
+
+  if (blogToDelete.user.toString() === requestUser._id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    return response.status(204).end()
+  }
+
+  response.status(401).json({ error: 'unauthorized' })
 })
 
 blogsRouter.put('/:id', async (request, response) => {
